t/io_uring: don't print partial IOPS etc output if exit signal was received
The run always terminates with what looks like a much slower cycle than
the previous seconds. That's not really the case, it's just that the
sleep() got interrupted by the signal and we slept less than we thought
we did, yet we still account it as a full second.

Just make it cleaner and break if finish is set.

Signed-off-by: Jens Axboe <axboe@kernel.dk>